requirements.txt
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
openai==1.42.0
python-dotenv==1.0.1
pypdf==4.3.1
docarray==0.40.0
pydantic==2.8.2


